requirements.txt
flask
psycopg2-binary
pandas
pyserial
sqlalchemy